# docs/swagger.yaml
basePath: /
definitions:
  dto.APIResponse:
    properties:
      data: {}
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
...
  dto.CreateVehicleRequest:
    properties:
      model:
        type: string
      number_plate:
        type: string
      organization_id:
        type: integer
      vin:
        type: string
    type: object
...
  dto.ErrorResponse:
    properties:
      message:
        type: string
      request_id:
        type: string
      status:
        type: string
    type: object
...
  dto.UpdateVehicleRequest:
    properties:
      number_plate:
        type: string
      organization_id:
        type: integer
      status:
        $ref: '#/definitions/model.VehicleStatus'
    type: object
...
  dto.VehicleResponse:
    properties:
      id:
        type: integer
      model:
        type: string
      number_plate:
        type: string
      organization_id:
        type: integer
      status:
        $ref: '#/definitions/model.VehicleStatus'
      vin:
        type: string
    type: object
...
  model.VehicleStatus:
    enum:
    - IDLE
    - ON_TRIP
    - IN_SERVICE
    - DELETED
    type: string
    x-enum-varnames:
    - VehicleStatusIdle
    - VehicleStatusOnTrip
    - VehicleStatusInService
    - VehicleStatusDeleted
host: localhost:8080
info:
  contact: {}
  description: Бэкенд системы мониторинга и телематики автопарка
  title: FleetTrack API
  version: "1.0"
paths:
...
  /vehicles:
    get:
      parameters:
      - description: VIN (17 символов)
        in: query
        name: vin
        type: string
      - description: Госномер
        in: query
        name: number_plate
        type: string
      - description: Модель
        in: query
        name: model
        type: string
      - description: Статус
        in: query
        name: status
        type: string
      - description: От даты создания (RFC3339)
        in: query
        name: created_from
        type: string
      - description: До даты создания (RFC3339)
        in: query
        name: created_to
        type: string
      - description: Лимит записей (по умолчанию 100, максимум 500)
        in: query
        name: limit
        type: integer
      - description: Смещение
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.VehicleResponse'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Список автомобилей организации
      tags:
      - vehicles
    post:
      consumes:
      - application/json
      parameters:
      - description: Данные автомобиля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateVehicleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Создать автомобиль
      tags:
      - vehicles
  /vehicles/{id}:
    delete:
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Удалить автомобиль
      tags:
      - vehicles
    get:
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Получить автомобиль по ID
      tags:
      - vehicles
    patch:
      consumes:
      - application/json
      parameters:
      - description: ID автомобиля
        in: path
        name: id
        required: true
        type: integer
      - description: Изменяемые поля
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateVehicleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.VehicleResponse'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: также возвращается, если в теле передан organization_id
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Изменить автомобиль
      tags:
      - vehicles
securityDefinitions:
  BearerAuth:
    description: 'Введите: Bearer <access_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"