# internal/handler/vehicle_handler.go
package handler

import (
	"context"
	"encoding/json"
	"fleettrack/internal/handler/dto"
	"fleettrack/internal/logger"
	"fleettrack/internal/middleware"
	"fleettrack/internal/model"
	"net/http"
	"strconv"

	"github.com/go-chi/chi/v5"
)

// VehicleHandler обрабатывает HTTP запросы, связанные с автомобилями
type VehicleHandler struct {
	vehicleService VehicleService
	logger         logger.Logger
}

// VehicleService определяет контракт бизнес-логики, необходимой VehicleHandler
type VehicleService interface {
	// CreateVehicle валидирует и сохраняет новый автомобиль
	CreateVehicle(ctx context.Context, v model.Vehicle) (model.Vehicle, error)
	// GetVehicleList возвращает список автомобилей по фильтру
	GetVehicleList(ctx context.Context, filter model.VehicleFilter) ([]model.Vehicle, error)
	// GetVehicleByID возвращает автомобиль по его ID
	GetVehicleByID(ctx context.Context, id int) (model.Vehicle, error)
	// DeleteVehicleByID удаляет автомобиль по его ID
	DeleteVehicleByID(ctx context.Context, id int, organizationID *int) (model.Vehicle, error)
	// UpdateVehicle
	UpdateVehicleByID(ctx context.Context, id int, upd model.UpdateVehicle, organizationID *int) (model.Vehicle, error)
}

// NewVehicleHandler создаёт новый VehicleHandler с переданными сервисом и логгером
func NewVehicleHandler(s VehicleService, l logger.Logger) *VehicleHandler {
	return &VehicleHandler{
		vehicleService: s,
		logger:         l,
	}
}

// HandlePostVehicle обрабатывает POST запрос на создание нового автомобиля
// @Summary Создать автомобиль
// @Tags vehicles
// @Accept json
// @Produce json
// @Param request body dto.CreateVehicleRequest true "Данные автомобиля"
// @Success 201 {object} dto.APIResponse{data=dto.VehicleResponse}
// @Failure 400 {object} dto.ErrorResponse
// @Failure 401 {object} dto.ErrorResponse
// @Failure 403 {object} dto.ErrorResponse
// @Failure 409 {object} dto.ErrorResponse
// @Router /vehicles [post]
// @Security BearerAuth
func (h *VehicleHandler) HandlePostVehicle(w http.ResponseWriter, r *http.Request) {
	var vehicleData dto.CreateVehicleRequest
	err := json.NewDecoder(r.Body).Decode(&vehicleData)
	if err != nil {
		respondError(w, r, h.logger, model.ErrInvalidJSON)
		return
	}

	authCtx, ok := middleware.AuthFromContext(r.Context())
	if !ok {
		respondError(w, r, h.logger, model.ErrMissingToken)
		return
	}

	vehicle := vehicleData.ToDomainModel()
	vehicle.OrganizationID = authCtx.OrganizationID
	savedVehicle, err := h.vehicleService.CreateVehicle(r.Context(), vehicle)
	if err != nil {
		respondError(w, r, h.logger, err)
		return
	}
	vehicleResponse := dto.VehicleResponse{
		ID:             savedVehicle.ID,
		OrganizationID: savedVehicle.OrganizationID,
		VIN:            savedVehicle.VIN,
		NumberPlate:    savedVehicle.NumberPlate,
		Model:          savedVehicle.Model,
		Status:         savedVehicle.Status,
	}

	respondCreated(w, r, "vehicle created", h.logger, vehicleResponse)
}

// HandleDeleteVehicle обрабатывает DELETE запрос на удаление автомобиля по ID
// @Summary Удалить автомобиль
// @Tags vehicles
// @Produce json
// @Param id path int true "ID автомобиля"
// @Success 200 {object} dto.APIResponse{data=dto.VehicleResponse}
// @Failure 400 {object} dto.ErrorResponse
// @Failure 401 {object} dto.ErrorResponse
// @Failure 403 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Router /vehicles/{id} [delete]
// @Security BearerAuth
func (h *VehicleHandler) HandleDeleteVehicle(w http.ResponseWriter, r *http.Request) {
	idStr := chi.URLParam(r, "id")
	id, err := strconv.Atoi(idStr)
	if err != nil {
		respondError(w, r, h.logger, model.ErrInvalidVehicleID)
		return
	}

	authCtx, ok := middleware.AuthFromContext(r.Context())
	if !ok {
		respondError(w, r, h.logger, model.ErrMissingToken)
		return
	}
	orgScope := scopeOrganizationID(authCtx)

	deletedVehicle, err := h.vehicleService.DeleteVehicleByID(r.Context(), id, orgScope)
	if err != nil {
		respondError(w, r, h.logger, err)
		return
	}

	vehicleResponse := dto.VehicleResponse{
		ID:             deletedVehicle.ID,
		OrganizationID: deletedVehicle.OrganizationID,
		VIN:            deletedVehicle.VIN,
		NumberPlate:    deletedVehicle.NumberPlate,
		Model:          deletedVehicle.Model,
		Status:         deletedVehicle.Status,
	}

	respondSuccess(w, r, "vehicle deleted", h.logger, vehicleResponse)
}

// HandleGetListVehicle возвращает список автомобилей с фильтрами
// @Summary Список автомобилей организации
// @Tags vehicles
// @Produce json
// @Param vin query string false "VIN (17 символов)"
// @Param number_plate query string false "Госномер"
// @Param model query string false "Модель"
// @Param status query string false "Статус"
// @Param created_from query string false "От даты создания (RFC3339)"
// @Param created_to query string false "До даты создания (RFC3339)"
// @Param limit query int false "Лимит записей (по умолчанию 100, максимум 500)"
// @Param offset query int false "Смещение"
// @Success 200 {object} dto.APIResponse{data=[]dto.VehicleResponse}
// @Failure 400 {object} dto.ErrorResponse
// @Failure 401 {object} dto.ErrorResponse
// @Router /vehicles [get]
// @Security BearerAuth
func (h *VehicleHandler) HandleGetListVehicle(w http.ResponseWriter, r *http.Request) {
	filter, err := dto.ParseVehicleFilter(r.URL.Query())
	if err != nil {
		respondError(w, r, h.logger, err)
		return
	}

	authCtx, ok := middleware.AuthFromContext(r.Context())
	if !ok {
		respondError(w, r, h.logger, model.ErrMissingToken)
		return
	}
	filter.OrganizationID = scopeOrganizationID(authCtx)

	vehicles, err := h.vehicleService.GetVehicleList(r.Context(), filter)
	if err != nil {
		respondError(w, r, h.logger, err)
		return
	}

	responses := make([]dto.VehicleResponse, 0, len(vehicles))
	for _, t := range vehicles {
		responses = append(responses, dto.VehicleResponse{
			ID:             t.ID,
			OrganizationID: t.OrganizationID,
			VIN:            t.VIN,
			NumberPlate:    t.NumberPlate,
			Model:          t.Model,
			Status:         t.Status,
		})
	}

	respondSuccess(w, r, "vehicles list", h.logger, responses)
}

// HandlePatchVehicle отвечает за изменение некоторых данных автомобиля
// @Summary Изменить автомобиль
// @Tags vehicles
// @Accept json
// @Produce json
// @Param id path int true "ID автомобиля"
// @Param request body dto.UpdateVehicleRequest true "Изменяемые поля"
// @Success 200 {object} dto.APIResponse{data=dto.VehicleResponse}
// @Failure 400 {object} dto.ErrorResponse
// @Failure 401 {object} dto.ErrorResponse
// @Failure 403 {object} dto.ErrorResponse "также возвращается, если в теле передан organization_id"
// @Failure 404 {object} dto.ErrorResponse
// @Router /vehicles/{id} [patch]
// @Security BearerAuth
func (h *VehicleHandler) HandlePatchVehicle(w http.ResponseWriter, r *http.Request) {
	id, err := strconv.Atoi(chi.URLParam(r, "id"))
	if err != nil {
		respondError(w, r, h.logger, model.ErrInvalidVehicleID)
		return
	}

	var request dto.UpdateVehicleRequest
	err = json.NewDecoder(r.Body).Decode(&request)
	if err != nil {
		respondError(w, r, h.logger, model.ErrInvalidJSON)
		return
	}

	authCtx, ok := middleware.AuthFromContext(r.Context())
	if !ok {
		respondError(w, r, h.logger, model.ErrMissingToken)
		return
	}
	if request.OrganizationID != nil {
		respondError(w, r, h.logger, model.ErrForbidden)
		return
	}
	orgScope := scopeOrganizationID(authCtx)

	upd := request.ToDomainModel()
	changed, err := h.vehicleService.UpdateVehicleByID(r.Context(), id, upd, orgScope)
	if err != nil {
		respondError(w, r, h.logger, err)
		return
	}

	vehicleResponse := dto.VehicleResponse{
		ID:             changed.ID,
		OrganizationID: changed.OrganizationID,
		VIN:            changed.VIN,
		NumberPlate:    changed.NumberPlate,
		Model:          changed.Model,
		Status:         changed.Status,
	}
	respondSuccess(w, r, "vehicle was changed", h.logger, vehicleResponse)
}

// HandleGetVehicleByID получает машину по ID
// @Summary Получить автомобиль по ID
// @Tags vehicles
// @Produce json
// @Param id path int true "ID автомобиля"
// @Success 200 {object} dto.APIResponse{data=dto.VehicleResponse}
// @Failure 400 {object} dto.ErrorResponse
// @Failure 401 {object} dto.ErrorResponse
// @Failure 404 {object} dto.ErrorResponse
// @Router /vehicles/{id} [get]
// @Security BearerAuth
func (h *VehicleHandler) HandleGetVehicleByID(w http.ResponseWriter, r *http.Request) {
	id, err := strconv.Atoi(chi.URLParam(r, "id"))
	if err != nil {
		respondError(w, r, h.logger, model.ErrInvalidVehicleID)
		return
	}

	authCtx, ok := middleware.AuthFromContext(r.Context())
	if !ok {
		respondError(w, r, h.logger, model.ErrMissingToken)
		return
	}

	vehicle, err := h.vehicleService.GetVehicleByID(r.Context(), id)
	if err != nil {
		respondError(w, r, h.logger, err)
		return
	}
	if err := requireOwnOrg(authCtx, vehicle.OrganizationID); err != nil {
		respondError(w, r, h.logger, err)
		return
	}

	vehicleResponse := dto.VehicleResponse{
		ID:             vehicle.ID,
		VIN:            vehicle.VIN,
		Model:          vehicle.Model,
		OrganizationID: vehicle.OrganizationID,
		NumberPlate:    vehicle.NumberPlate,
		Status:         vehicle.Status,
	}

	respondSuccess(w, r, "vehicle found", h.logger, vehicleResponse)
}